% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{calculate_classification_metrics}
\alias{calculate_classification_metrics}
\title{Calculate Classification Metrics}
\usage{
calculate_classification_metrics(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\value{
A list containing True Positive Rate (TPR), False Positive Rate (FPR), and Precision
}
\description{
This function computes classification metrics including True Positive Rate (TPR), False Positive Rate (FPR), and Precision for a binary classification model.
}
\examples{
calculate_classification_metrics(unceRtain_object)

}
