% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{calculate_EPE}
\alias{calculate_EPE}
\title{Calculate Expected Prediction Error (EPE)}
\usage{
calculate_EPE(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\value{
The calculated EPE
}
\description{
This function calculates the Expected Prediction Error (EPE) for a binary classification model.
}
\examples{
calculate_EPE(unceRtain_object)

}
