% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{calculate_misclassification_cost}
\alias{calculate_misclassification_cost}
\title{Calculate Mis-Classification Cost and Misclassification Error}
\usage{
calculate_misclassification_cost(unceRtain_object, cfp, cfn)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}

\item{cfp}{Cost of false positive predictions}

\item{cfn}{Cost of false negative predictions}
}
\value{
A list containing Misclassification Cost and Misclassification Error
}
\description{
This function computes the mis-classification cost and misclassification error for a binary classification model.
}
\examples{
cfp <- 1  # Cost of False Positive
cfn <- 2  # Cost of False Negative
calculate_misclassification_cost(unceRtain_object, cfp, cfn)

}
