% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{confusion_matrix_metrics_plot}
\alias{confusion_matrix_metrics_plot}
\title{Calculate Confusion Matrix Metrics}
\usage{
confusion_matrix_metrics_plot(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\value{
A table representing the confusion matrix
}
\description{
This function calculates the confusion matrix for a binary classification model and returns the result.
}
\examples{
confusion_matrix_metrics_plot(unceRtain_object)

}
