% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.analysis.filenames.r
\name{ensemble.filename}
\alias{ensemble.filename}
\title{Generate ensemble filenames}
\usage{
ensemble.filename(
  settings,
  prefix = "ensemble.samples",
  suffix = "Rdata",
  all.var.yr = TRUE,
  ensemble.id = settings$ensemble$ensemble.id,
  variable = settings$ensemble$variable,
  start.year = settings$ensemble$start.year,
  end.year = settings$ensemble$end.year
)
}
\arguments{
\item{settings}{list of PEcAn settings.}

\item{prefix}{string to appear at the beginning of the filename}

\item{suffix}{file extension: string to appear at the end of the filename}

\item{all.var.yr}{logical: does ensemble include all vars and years?
If FALSE, filename will include years and vars}

\item{ensemble.id}{ensemble ID(s)}

\item{variable}{variable(s) included in the ensemble.}

\item{start.year, end.year}{first and last year simulated.}
}
\value{
a vector of filenames, each in the form
  `[settings$outdir]/[prefix].[ensemble.ID].[variable].[start.year].[end.year][suffix]`.
}
\description{
Generates a vector of filenames to be used for PEcAn ensemble output files.
All paths start from directory `settings$outdir`,
which will be created if it does not exist.
}
\details{
Typically used by passing only a settings object,
  but all values can be overridden for manual use.

If only a single variable or a subset of years are needed,
  the generated filename will identify these in the form
If all vars and years are included, set `all.yr.var` to TRUE
  to get a filename of the form `prefix.ensemble_id.suffix`.
All elements are recycled vectorwise.
}
\author{
Ryan Kelly
}
