% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_states}
\alias{ensemble_states}
\title{Function for preparing and writing ensemble states}
\usage{
ensemble_states(settings, ensemble.samples, con)
}
\arguments{
\item{settings}{list of PEcAn settings}

\item{ensemble.samples}{list of lists supplied by \link{get.ensemble.samples}}

\item{con}{DB connection}
}
\value{
samples
}
\description{
Function for preparing and writing ensemble states
}
\examples{
\dontrun{ensemble_states(settings, con)}

}
