% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{find_optimal_threshold}
\alias{find_optimal_threshold}
\title{Find Optimal Threshold for Classification}
\usage{
find_optimal_threshold(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\value{
The optimal threshold for classification
}
\description{
This function determines the optimal threshold for classification based on the ROC curve.
}
\examples{
find_optimal_threshold(unceRtain_object)

}
