% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{generate_confusion_matrix_visualization}
\alias{generate_confusion_matrix_visualization}
\title{Generate Confusion Matrix Visualization}
\usage{
generate_confusion_matrix_visualization(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\description{
This function generates a visualization for the confusion matrix.
}
\examples{
generate_confusion_matrix_visualization(unceRtain_object)

}
