% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{generate_roc}
\alias{generate_roc}
\title{Generate ROC Curve for Model Evaluation}
\usage{
generate_roc(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\description{
This function generates the Receiver Operating Characteristic (ROC) curve, providing insights into the binary classification model's performance.
}
\examples{
generate_roc(unceRtain_object)

}
