% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{generate_roc_curve_visualization}
\alias{generate_roc_curve_visualization}
\title{Generate ROC Curve Visualization}
\usage{
generate_roc_curve_visualization(unceRtain_object)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}
}
\description{
This function generates a visualization for the ROC curve.
}
\examples{
generate_roc_curve_visualization(unceRtain_object)

}
