% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.gi.phii}
\alias{get.gi.phii}
\title{Get g_i(phi_i)}
\usage{
get.gi.phii(splinefuns, trait.samples, maxn = NULL)
}
\arguments{
\item{splinefuns}{univariate spline functions created for each trait, e.g. by the \code{\link{sensitivity.analysis}} function.}

\item{trait.samples}{n x m matrix (or list with m vectors of length n) of n parameter sets, each with a sample from m traits}

\item{maxn}{maximum number of parameter sets to evaluate}
}
\value{
matrix of spline estimates of model output for each of n parameter sets
}
\description{
Calculate distribution of function of a variable
}
\details{
Transform trait parameter through trait-specific univariate model emulator spline function
Given a matrix of parameter sets, calculates a matrix of model output by applying appropriate spline function to each parameter. Output is use with \code{\link{variance.decomposition}} and \code{\link{spline.ensemble}}
}
\author{
David LeBauer
}

