% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sa_sample_list.R
\name{get_sa_sample_list}
\alias{get_sa_sample_list}
\title{get sensitivity samples as a list}
\usage{
get_sa_sample_list(pft, env, quantiles)
}
\arguments{
\item{pft}{Plant Functional Type}

\item{env}{}

\item{quantiles}{quantiles at which to obtain samples from parameter for
sensitivity analysis}
}
\description{
get sensitivity samples as a list
}
