% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{input.ens.gen}
\alias{input.ens.gen}
\title{Function for generating samples based on sampling method, parent or etc}
\usage{
input.ens.gen(
  settings,
  ensemble_size,
  input,
  method = "sampling",
  parent_ids = NULL
)
}
\arguments{
\item{settings}{list of PEcAn settings}

\item{ensemble_size}{size of ensemble}

\item{input}{name of input to sample, e.g. "met", "veg", "pss"}

\item{method}{Method for sampling - For now looping or sampling with replacement is implemented}

\item{parent_ids}{This is basically the order of the paths that the parent is sampled.See Details.}
}
\value{
For a given input/tag in the pecan xml and a method, this function returns a list with $id showing the order of sampling and $samples with samples of that input.
}
\description{
Function for generating samples based on sampling method, parent or etc
}
\details{
If for example met was a parent and it's sampling method resulted in choosing the first, third and fourth samples, these are the ids that need to be sent as
parent_ids to this function.
}
\examples{
\dontrun{
input.ens.gen(settings, "met", "sampling")
}

}
