% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{MakePredictions}
\alias{MakePredictions}
\title{Make Predictions using Logistic Regression Model}
\usage{
MakePredictions(model, test_data, observed_outcome)
}
\arguments{
\item{model}{A logistic regression model object.}

\item{test_data}{A data frame containing predictor variables for making predictions.}

\item{observed_outcome}{A numeric vector or data frame representing the observed outcomes.}
}
\value{
A list containing predictions and observed outcomes.
}
\description{
This function takes a logistic regression model, test data, and observed outcomes as input,
checks for compatibility, and returns predictions along with observed outcomes.
}
\examples{
# Example with logistic regression model, test data, and observed outcomes
model <- glm(formula = Outcome ~ Predictor1 + Predictor2, data = training_data, family = binomial)
test_data <- data.frame(Predictor1 = c(1, 2, 3), Predictor2 = c(4, 5, 6))
observed_outcome <- c(0, 1, 0)
predictions <- MakePredictions(model, test_data, observed_outcome)

}
\seealso{
\code{\link{predict}}

\code{\link{glm}}

\code{\link{is_predictable}}
}
