\name{plot.variance.decomposition}
\alias{plot.variance.decomposition}
\title{Variance Decomposition Plots}
\usage{
  plot.variance.decomposition(plot.inputs, outdir,
    prior.plot.inputs = NULL,
    fontsize = list(title = 18, axis = 14), filter = TRUE,
    variance.scale = sqrt, variance.prefix = "Root")
}
\arguments{
  \item{plot.inputs}{}

  \item{outdir}{}

  \item{prior.plot.inputs}{}

  \item{fontsize}{}

  \item{filter}{}

  \item{variance.scale}{}

  \item{variance.prefix}{}
}
\description{
  Plot results of variance decomposition
}
\details{
  Plots variance decomposition tryptich
}
\author{
  David LeBauer
}

