% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_uncertainty.R
\name{plot_flux_uncertainty}
\alias{plot_flux_uncertainty}
\title{Plot fit for heteroskedastic flux uncertainty}
\usage{
plot_flux_uncertainty(f, ...)
}
\arguments{
\item{f}{output of flux.uncertainty functions}

\item{...}{optional graphical paramters}
}
\description{
Plot fit for heteroskedastic flux uncertainty
}
\author{
Mike Dietze, Carl Davidson
}
