% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Sensitivity plot}
\usage{
plot_sensitivity(
  sa.sample,
  sa.spline,
  trait,
  y.range = c(0, 50),
  median.i = 4,
  prior.sa.sample = NULL,
  prior.sa.spline = NULL,
  fontsize = list(title = 12, axis = 8),
  linesize = 1,
  dotsize = 2
)
}
\arguments{
\item{sa.sample}{trait quantiles used in sensitivity analysis}

\item{sa.spline}{spline function estimated from sensitivity analysis}

\item{trait}{trait name for title}

\item{y.range}{limits for y axis of plot}

\item{median.i}{index of median value in sa.sample; \code{median.i == which(as.numeric(rownames(sa.sample)) == 50) }}

\item{prior.sa.sample}{similar to sa.sample, but for prior distribution. If given, plots sensitivity for prior run}

\item{prior.sa.spline}{similar to sa.spline, but for prior trait distribution.}

\item{fontsize}{(optional) list with three arguments that can be set to vary the fontsize of the title, axis labels, and axis title in the sensitivity plots}

\item{linesize}{passed to ggplot to set line thickness}

\item{dotsize}{passed to ggplot to set point size}
}
\value{
object of class ggplot
}
\description{
Plot univariate response of model output to a trait parameter.
}
\details{
Plots for a single trait; called by \code{\link{plot_sensitivities}}
to plot sensitivity plots for multiple traits.
}
