% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_variance_decomposition}
\alias{plot_variance_decomposition}
\title{Variance Decomposition Plots}
\usage{
plot_variance_decomposition(
  plot.inputs,
  fontsize = list(title = 18, axis = 14)
)
}
\arguments{
\item{plot.inputs}{Output from a sensitivity analysis. Output must be of the form
given by sensitivity.results$variance.decomposition.output in model output}

\item{fontsize}{list specifying the font size of the titles and axes of the graph}
}
\description{
Plots variance decomposition tryptich
}
\examples{
x <- list(trait.labels = c('a', 'b', 'c'),
          coef.vars = c(a=1,b=0.5, c=0.1),
          elasticities = c(a=1,b=2,c=0.5),
          variances    = c(a = 20, b=30, c = 10))
do.call(grid.arrange, c(plot_variance_decomposition(x), ncol = 4))
}
\author{
David LeBauer, Carl Davidson
}
