% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.data.assim.R
\name{prep.data.assim}
\alias{prep.data.assim}
\title{prep.data.assim}
\usage{
prep.data.assim(settings)
}
\arguments{
\item{settings}{the PEcAn settings object}
}
\value{
None
}
\description{
prep.data.assim
}
\section{purpose}{

This function sets up a call to PEcAn.assim.sequential::sda.enkf()
}

\author{
Luke Dramko
}
