% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processModelOutput.R
\name{processModelOutput}
\alias{processModelOutput}
\title{process model output and return the results as a list}
\usage{
processModelOutput()
}
\value{
an list of model outputs
}
\description{
process model output and return the results as a list
}
