% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{run.ensemble.analysis}
\alias{run.ensemble.analysis}
\title{run ensemble.analysis}
\usage{
run.ensemble.analysis(
  settings,
  plot.timeseries = NA,
  ensemble.id = NULL,
  variable = NULL,
  start.year = NULL,
  end.year = NULL,
  sobolSA = NULL,
  ...
)
}
\arguments{
\item{plot.timeseries}{if TRUE plots a modeled timeseries of target variable(s) with CIs}
}
\value{
nothing, creates ensemble plots as ensemble.analysis.pdf
}
\description{
run ensemble.analysis
}
\author{
David LeBauer, Shawn Serbin, Ryan Kelly
}
