% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.results.R
\name{runModule.get.results}
\alias{runModule.get.results}
\title{Apply get.results to each of a list of settings}
\usage{
runModule.get.results(settings)
}
\arguments{
\item{settings}{a PEcAn \code{Settings} or \code{MultiSettings} object}
}
\description{
Apply get.results to each of a list of settings
}
\seealso{
get.results
}
