% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{runModule.run.ensemble.analysis}
\alias{runModule.run.ensemble.analysis}
\title{Run ensemble analyses across all sites in settings}
\usage{
runModule.run.ensemble.analysis(settings, ...)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{...}{arguments passed on to run.ensemble.analysis}
}
\description{
Run ensemble analyses across all sites in settings
}
