% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.sensitivity.analysis.R
\name{runModule.run.sensitivity.analysis}
\alias{runModule.run.sensitivity.analysis}
\title{Run sensitivity analysis for every Settings in a MultiSettings}
\usage{
runModule.run.sensitivity.analysis(settings, ...)
}
\arguments{
\item{settings}{PEcan settings object}

\item{...}{additional arguments passed on to `run.sensitivity.analysis`}
}
\description{
Caution: Not yet working for multisite settings.
  It will _run_ the analysis for all sites,
  but each site will overwrite the result from the previous one.
}
