% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.analysis.R
\name{sa.splinefun}
\alias{sa.splinefun}
\title{Spline estimate of univariate relationship between parameter value and model output}
\usage{
sa.splinefun(quantiles.input, quantiles.output)
}
\arguments{
\item{quantiles.input}{passed to `x` argument of `stats::splinefun`}

\item{quantiles.output}{passed to `y` argument of `stats::splinefun`}
}
\value{
function
}
\description{
Creates a spline function using the splinefun function that estimates univariate response of parameter input to model output
}
