% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Script.R
\name{set_cost}
\alias{set_cost}
\title{Set Cost of False Positive and False Negative}
\usage{
set_cost(unceRtain_object, cfp, cfn)
}
\arguments{
\item{unceRtain_object}{An object created using the MakePrediction() function}

\item{cfp}{Cost of false positive predictions}

\item{cfn}{Cost of false negative predictions}
}
\value{
The total cost based on the provided false positive and false negative costs
}
\description{
This function allows you to set the cost associated with false positive and false negative predictions.
}
\examples{
set_cost(unceRtain_object)

}
