% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variance.R
\name{spline.ensemble}
\alias{spline.ensemble}
\title{Spline Ensemble}
\usage{
spline.ensemble(gi.phii, median)
}
\arguments{
\item{gi.phii}{matrix given as output from \code{\link{get.gi.phii}}}

\item{median}{median value around which variance will be calculated}
}
\description{
Estimate model output based on univariate splines
}
\details{
Accepts output from get.gi.phii (the matrix $g(\phi_i)$) and produces
spline estimate of $f(phi)$ for use in estimating closure term associated with
spline approximation
}
\author{
David LeBauer

David LeBauer
}

