% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varkernelslicerange.R
\name{varkernelslicerange}
\alias{varkernelslicerange}
\title{Estimated output variable values given the expected values of the influencing variable, 
based on a slice of 'z' from the Kernel density plot of the influencing variable and output 
variable data.}
\usage{
varkernelslicerange(in_var, out_var, min_in_var, max_in_var)
}
\arguments{
\item{in_var}{is a vector of observations of a given influencing variable corresponding to another list with observed values of an outcome variable {out_var}.}

\item{out_var}{is a vector of observed values of an outcome variable corresponding to another list with observations of a given influencing variable {in_var}.}

\item{min_in_var}{is a value of the lowest expected amount of {in_var} for which the outcome variable {out_var} should be estimated (must be < {max_in_var}).}

\item{max_in_var}{is a value of the highest expected amount of a given influencing variable {in_var} for which the outcome variable {out_var} should be estimated (must be > {min_in_var}).}
}
\description{
Plot representing probabilities (shown along the y-axis) for the expected outcome variable (shown along the x-axis). 
This is a broad slice through the density kernel from uncertainty::varkernel() function, which integrates to 1, the probability values are relative, not absolute measures.
}
\examples{
variable <- sample(x = 1:50, size = 20, replace = TRUE)
out_var <- sample(x = 1000:5000, size = 20, replace = TRUE)
varkernelslicerange(variable, out_var, 10, 20)

}
\keyword{density}
\keyword{influence}
\keyword{kernel}
