% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_based_sobolSA.R
\name{cluster_based_sobolSA}
\alias{cluster_based_sobolSA}
\title{Reads an ensemble time-series from PEcAn for the selected target variable}
\usage{
cluster_based_sobolSA(
  settings,
  ensemble.id = NULL,
  variable = "NPP",
  start.year = NULL,
  end.year = NULL
)
}
\value{
list
}
\description{
Performs cluster-based GSA based on Roux et al 2021
function uses code from https://github.com/sbuis/ClusterBased_GSA/blob/master/clusterBasedGSA_on_Toycurve.Rmd
}
