% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sobol_indices.R
\name{compute_sobol_indices}
\alias{compute_sobol_indices}
\title{Compute Sobol indices from a finished PEcAn run}
\usage{
compute_sobol_indices(outdir, sobol_obj, var = "GPP", stat_fun = mean)
}
\arguments{
\item{outdir}{PEcAn run output directory that contains runs.txt}

\item{sobol_obj}{object produced by PEcAn.uncertainty::generate_joint_ensemble_design()}

\item{var}{Variable name to summarise (default "GPP").}

\item{stat_fun}{Summary statistic applied to var default mean .}
}
\value{
sobol_obj 
.
}
\description{
Loads model outputs from a Sobol ensemble, calculates summary
statistics for a chosen variable, feeds them to \code{sensitivity::tell()},
and returns the updated Sobol object.
}
