% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{ensemble.ts}
\alias{ensemble.ts}
\title{Plots an ensemble time-series from PEcAn for the selected target variable}
\usage{
ensemble.ts(ensemble.ts, observations = NULL, window = 1, ...)
}
\arguments{
\item{ensemble.ts}{ensemble timeseries to be plotted, as returned by [read.ensemble.ts()]}

\item{observations}{observed data to plot over timeseries predictions}

\item{window}{number of timepoints to average across}

\item{...}{additional arguments, currently ignored}
}
\value{
nothing, generates an ensemble time-series plot
}
\description{
Plots an ensemble time-series from PEcAn for the selected target variable
}
\author{
Michael Dietze, Ryan Kelly
}
