% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{ensemble.ts}
\alias{ensemble.ts}
\title{Plots an ensemble time-series from PEcAn for the selected target variable}
\usage{
ensemble.ts(ensemble.ts, observations = NULL, window = 1)
}
\value{
nothing, generates an ensemble time-series plot
}
\description{
Plots an ensemble time-series from PEcAn for the selected target variable
}
\author{
Michael Dietze, Ryan Kelly
}

