% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ensemble.ts}
\alias{ensemble.ts}
\title{Plots an ensemble time-series from PEcAn for the selected target variables}
\usage{
ensemble.ts(ensemble.ts, observations = NULL, window = 1)
}
\value{
nothing, generates an ensemble time-series plot
}
\description{
Plots an ensemble time-series from PEcAn for the selected target variables
}
\author{
Michael Dietze
}

