% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flux_uncertainty.R
\name{get.change}
\alias{get.change}
\title{Get delta between sequential flux datapoints}
\usage{
get.change(measurement)
}
\value{
Difference between consecutive measurements
}
\description{
Get delta between sequential flux datapoints
}
\author{
Mike Dietze, Carl Davidson
}

