\name{get.coef.var}
\alias{get.coef.var}
\title{Get coefficient of variance}
\usage{
  get.coef.var(set)
}
\arguments{
  \item{set}{numeric vector of trait values}
}
\value{
  coeficient of variance
}
\description{
  Given a set of numbers (a numeric vector), this returns
  the set's coefficient of variance.
}

