% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parameter.samples.R
\name{get.parameter.samples}
\alias{get.parameter.samples}
\title{Convert priors / MCMC samples to chains that can be sampled for model parameters}
\usage{
get.parameter.samples(
  settings,
  posterior.files = rep(NA, length(settings$pfts)),
  ens.sample.method = "uniform"
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{posterior.files}{list of filenames to read from}

\item{ens.sample.method}{one of "halton", "sobol", "torus", "lhc", "uniform"}
}
\description{
Convert priors / MCMC samples to chains that can be sampled for model parameters
}
\author{
David LeBauer, Shawn Serbin, Istem Fer
}
