% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensitivities}
\alias{plot.sensitivities}
\title{Plot Sensitivities}
\usage{
\method{plot}{sensitivities}(sensitivity.plot.inputs,
  prior.sensitivity.plot.inputs = NULL, ...)
}
\arguments{
\item{sensitivity.plot.inputs}{inputs}

\item{prior.sensitivity.plot.inputs}{priors}

\item{...}{arguments passed to \code{\link{plot.sensitivity}}}

\item{sensitivity.results}{list containing sa.samples and sa.splines}
}
\value{
list of plots, one per trait
}
\description{
Plot functions and quantiles used in sensitivity analysis
}
\details{
Generates a plot using \code{\link{plot.sensitivity}} for multiple traits.
}

