% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_uncertainty.R
\name{read.ameriflux.L2}
\alias{read.ameriflux.L2}
\title{Read Ameriflux L2 Data}
\usage{
read.ameriflux.L2(file.name, year)
}
\value{
Ameriflux L2 data read from file
}
\description{
Read Ameriflux L2 Data
}
\author{
Mike Dietze, Carl Davidson
}

