% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.ensemble.ts}
\alias{read.ensemble.ts}
\title{Reads an ensemble time-series from PEcAn for the selected target variable}
\usage{
read.ensemble.ts(model, ensemble.id = NULL, variable = NULL,
  start.year = NULL, end.year = NULL)
}
\value{
list
}
\description{
Reads ensemble time-series from PEcAn for the selected target variable
}
\author{
Michael Dietze, Ryan Kelly
}

