% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{read.ensemble.ts}
\alias{read.ensemble.ts}
\title{Reads ensemble time-series from PEcAn for the selected target variable}
\usage{
read.ensemble.ts(
  settings,
  ensemble.id = NULL,
  variable = NULL,
  start.year = NULL,
  end.year = NULL
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{ensemble.id}{database ID, taken from settings if not specified}

\item{variable}{variable name to process. Required}

\item{start.year, end.year}{taken from settings if not specified}
}
\value{
list
}
\description{
Reads ensemble time-series from PEcAn for the selected target variable
}
\author{
Michael Dietze, Ryan Kelly
}
