% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{read.ensemble.ts}
\alias{read.ensemble.ts}
\title{Reads an ensemble time-series from PEcAn for the selected target variables}
\usage{
read.ensemble.ts(model)
}
\value{
list
}
\description{
Reads ensemble time-series from PEcAn for the selected target variables
}
\author{
Michael Dietze
}

