% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.ensemble.ts}
\alias{read.ensemble.ts}
\title{Reads an ensemble time-series from PEcAn for the selected target variables}
\usage{
read.ensemble.ts(model)
}
\value{
list
}
\description{
Reads ensemble time-series from PEcAn for the selected target variables
}
\author{
Michael Dietze
}

