% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ensemble.analysis.R
\name{run.ensemble.analysis}
\alias{run.ensemble.analysis}
\title{Run ensemble analysis and generate summary plots}
\usage{
run.ensemble.analysis(
  settings,
  plot.timeseries = NA,
  ensemble.id = NULL,
  variable = NULL,
  start.year = NULL,
  end.year = NULL,
  ...
)
}
\arguments{
\item{settings}{PEcAn \code{Settings} object}

\item{plot.timeseries}{if TRUE plots a modeled timeseries of target variable(s) with CIs}

\item{ensemble.id}{database ID, taken from settings if not specified}

\item{variable}{variable name(s) to process, taken from settings if not specified
taken as a vector in case of multiple variables}

\item{start.year}{First year of analysis, taken from settings if not specified}

\item{end.year}{Last year of analysis, taken from settings if not specified}

\item{...}{additional arguments passed to [ensemble.ts()]}
}
\value{
nothing, creates ensemble plots as ensemble.analysis.pdf.
  If /code{plot.timeseries = TRUE}, creates ensemble.ts.pdf and 
  also saves ts analysis object as \code{.Rdata} file
}
\description{
Generates distribution and optional time-series plots for ensemble
model outputs. Reads processed ensemble output files and produces
summary PDFs in \code{settings$outdir}.
}
\details{
Produces histograms, boxplots, and optionally time-series plots with
ensemble mean and 95% confidence intervals. Requires ensemble output
files generated by \code{\link{get.results}}
}
\seealso{
\code{\link{get.results}}, \code{\link{read.ensemble.ts}}, \code{\link{ensemble.ts}}
}
\author{
David LeBauer, Shawn Serbin, Ryan Kelly
}
