\name{run.ensemble.analysis}
\alias{run.ensemble.analysis}
\title{run ensemble.analysis}
\usage{
  run.ensemble.analysis(plot.timeseries = NA)
}
\arguments{
  \item{plot.timeseries}{if TRUE plots a modeled timeseries
  of target variable(s) with CIs}
}
\value{
  nothing, creates ensemble plots as ensemble.analysis.pdf
}
\description{
  run ensemble.analysis
}
\author{
  David LeBauer, Shawn Serbin
}

