\name{run.sensitivity.analysis}
\alias{run.sensitivity.analysis}
\title{run sensitivity.analysis}
\usage{
  run.sensitivity.analysis()
}
\value{
  nothing, saves \code{sensitivity.results} as
  sensitivity.results.Rdata, sensitivity plots as
  sensitivityanalysis.pdf, and variance decomposition
  'popsicle plot' as variancedecomposition.pdf a side
  effect.
}
\description{
  run sensitivity.analysis
}
\author{
  David LeBauer
}

