% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.sensitivity.analysis}
\alias{run.sensitivity.analysis}
\title{run sensitivity.analysis}
\usage{
run.sensitivity.analysis(plot = TRUE)
}
\arguments{
\item{plot}{logical. Option to generate sensitivity analysis and variance
decomposition plots (plot=TRUE) or to turn these plots off (plot=FALSE).}
}
\value{
nothing, saves \code{sensitivity.results} as sensitivity.results.Rdata,
sensitivity plots as sensitivityanalysis.pdf, and variance decomposition 'popsicle plot'
as variancedecomposition.pdf a side effect (OPTIONAL)
}
\description{
run sensitivity.analysis
}
\author{
David LeBauer, Shawn Serbin
}

