% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensitivity.analysis.R
\name{sa.splinefun}
\alias{sa.splinefun}
\title{Sensitivity spline function}
\usage{
sa.splinefun(quantiles.input, quantiles.output)
}
\arguments{
\item{quantiles.input}{}

\item{quantiles.output}{}
}
\value{
function
}
\description{
Spline estimate of univariate relationship between parameter value and model output
}
\details{
Creates a spline function using the splinefun function that estimates univariate response of parameter input to model output
}

