% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sa.splinefun}
\alias{sa.splinefun}
\title{Sensitivity spline function}
\usage{
sa.splinefun(quantiles.input, quantiles.output)
}
\arguments{
\item{quantiles.input}{}

\item{quantiles.output}{}
}
\value{
function
}
\description{
Spline estimate of univariate relationship between parameter value and model output
}
\details{
Creates a spline function using the splinefun function that estimates univariate response of parameter input to model output
}

